Run generate before bundle
The previous code was triggering the module generation after the
initial bundle. It did that to load the configuration in package.json
like elwin013/parcel-plugin-static-files-copy does.

The problem is that this approach doesn't work for this plugin, because
we need generate the svg module before Elm compiles our app.
Otherwise we would get the previous version of our assets.

This was not a problem in development because when you change your assets
it triggers a build, however when bundling the production build it does not.
I didn't catch this before because in my testing we only run the production
build in CI and the develpment build updated the elm files before I
pushed.

This will be released in 1.0.2.